consistent use of int64 for ids (was int32)

+++ test_utils/test_utils.go
@@ -17,11 +17,11 @@ func GetApiToken() (string, error) {
 	return "", fmt.Errorf("%s env var not specified", ENV_LOGZIO_API_TOKEN)
 }
 
-func GetAccountId() (int32, error) {
+func GetAccountId() (int64, error) {
 	account_id_string := os.Getenv(ENV_LOGZIO_ACCOUNT_ID)
 	if len(account_id_string) == 0 {
 		return -1, fmt.Errorf("%s env var not specified", ENV_LOGZIO_ACCOUNT_ID)
 	}
 	account_id, _ := strconv.ParseInt(account_id_string, 10, 32)
-	return int32(account_id), nil
+	return int64(account_id), nil
 }
